# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Menu Engineering API",
        "description": "API para gestão de engenharia de cardápio",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "46448289940e329a644b891be9c66943",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "494913cea16e9bd1ed0d3a25eda1b1f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "61665c36ed1115a4d53992f26035a8fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "ff2f374ba56017e8d111c4e218868b9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/recipe-items": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "4362e63fea325bf5ef83dcb0962edf09",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "dddb01f83a22bb0e6703150059183db1",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
        "/api/recipe-items/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "b13cf5f3e190cf1b8abd992629a49b62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "343942a595ec1d47131a280727726ff2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "7b396d7355c476108ee47eddd22a6f19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    }
                }
            }
        },
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "operationId": "419ebb078c83262b05f7c639ae8b2c21",
                "responses": {
                    "200": {
                        "description": "Lista"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Cria venda e baixa estoque por FIFO",
                "operationId": "51f42c056032d143981f26f6c0888f24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sold_at",
                                    "items"
                                ],
                                "properties": {
                                    "sold_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "channel": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "dish_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "422": {
                        "description": "Erro de validação / estoque insuficiente"
                    }
                }
            }
        },
        "/api/sales/{id}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "operationId": "b352670063e2f09edbfb74f39aef9b56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe"
                    }
                }
            }
        },
...
    },
...
}